basic shape

+++ src/core/shape.cpp
@@ -1,5 +1,9 @@
 #include "core/shape.h"
 #include "core/interaction.h"
+#include "core/bounding_boxes.h"
+#include "core/transform.h"
+
+
 
 Shape::Shape(const Transform* ObjectToWorld, const Transform* WorldToObject, bool reverseOrientation) : 
 	ObjectToWorld(ObjectToWorld),
@@ -7,13 +11,15 @@ Shape::Shape(const Transform* ObjectToWorld, const Transform* WorldToObject, boo
 	reverseOrientation(reverseOrientation),
 	transformSwapsHandedness(ObjectToWorld->SwapsHandedness()) {}
 
-Bounds3f Shape::WorldBound() const {
+Bounds3f Shape::WorldBound() const
+{
 	return (*ObjectToWorld)(ObjectBound());
 }
 
 bool Shape::IntersectP(const Ray& ray, bool testAlphaTexture) const
 {
-	double tHit = ray.tMax;
+	Float tHit = ray.tMax;
 	SurfaceInteraction isect;
 	return Intersect(ray, tHit, &isect, testAlphaTexture);
 }
+
